requirements.txt
# Core
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0

# Real-Time Retrieval
faiss-cpu>=1.7.4  # or faiss-gpu if using GPU
# redis>=4.5.0    # optional if using Redis instead of Faiss

# Kafka for streaming
kafka-python>=2.0.2

# MLOps
mlflow>=2.4.1
dvc[s3]>=3.20.0    # Adjust for your storage backend: [s3], [gdrive], etc.

# General
matplotlib>=3.7.0
seaborn>=0.12.2
notebook
ipykernel


uvicorn[standard]
